feat(core): add plugin API
