+++ packages/core/src/composables/useVueFlow.ts
@@ -1,85 +1,7 @@
 import type { EffectScope } from 'vue'
-import type { FlowOptions, FlowProps, State, VueFlowStore } from '~/types'
+import type { FlowProps, VueFlowStore } from '~/types'
 import { VueFlow } from '~/context'
-import { useActions, useGetters, useState } from '~/store'
-
-/**
- * Stores all currently created store instances
- */
-export class Storage {
-  public currentId = 0
-  public flows = new Map<string, VueFlowStore>()
-  static instance: Storage
-
-  public static getInstance(): Storage {
-    if (!Storage.instance) {
-      Storage.instance = new Storage()
-    }
-
-    return Storage.instance
-  }
-
-  public set(id: string, flow: VueFlowStore) {
-    return this.flows.set(id, flow)
-  }
-
-  public get(id: string) {
-    return this.flows.get(id)
-  }
-
-  public remove(id: string) {
-    return this.flows.delete(id)
-  }
-
-  public create(id: string, preloadedState?: FlowOptions): VueFlowStore {
-    const state: State = useState(preloadedState)
-
-    const reactiveState = reactive(state)
-
-    const getters = useGetters(reactiveState)
-
-    const actions = useActions(reactiveState, getters)
-
-    const hooksOn = <any>{}
-    Object.entries(reactiveState.hooks).forEach(([n, h]) => {
-      const name = `on${n.charAt(0).toUpperCase() + n.slice(1)}`
-      hooksOn[name] = h.on
-    })
-
-    const emits = <any>{}
-    Object.entries(reactiveState.hooks).forEach(([n, h]) => {
-      emits[n] = h.trigger
-    })
-
-    actions.setState(reactiveState)
-
-    if (preloadedState) {
-      if (preloadedState.modelValue) actions.setElements(preloadedState.modelValue)
-      if (preloadedState.nodes) actions.setNodes(preloadedState.nodes)
-      if (preloadedState.edges) actions.setEdges(preloadedState.edges)
-    }
-
-    const flow: VueFlowStore = {
-      ...hooksOn,
-      ...getters,
-      ...actions,
-      ...toRefs(reactiveState),
-      emits,
-      id,
-      $destroy: () => {
-        this.remove(id)
-      },
-    }
-
-    this.set(id, flow)
-
-    return flow
-  }
-
-  public getId() {
-    return `vue-flow-${this.currentId++}`
-  }
-}
+import { Storage } from '~/storage'
 
 type Injection = VueFlowStore | null | undefined
 type Scope = (EffectScope & { vueFlowId: string }) | undefined
