+++ packages/vue-flow/src/index.ts
@@ -1,3 +1,8 @@
+/**
+ * Vue Flow
+ * @module vue-flow
+ */
+
 import '@vue-flow/core/dist/style.css'
 import '@vue-flow/core/dist/theme-default.css'
 export * from '@vue-flow/additional-components'
